refactor: Improve endpoint responses with ResponseEntity

+++ src/main/java/Ant0_n10/Java10x/CadastroDeNinjas/Ninjas/NinjaService.java
@@ -20,31 +20,39 @@ public class NinjaService {
     public NinjaDTO criar(NinjaDTO ninjaDTO) {
         NinjaModel ninjaModel = ninjaMapper.map(ninjaDTO);
         ninjaModel = ninjaRepository.save(ninjaModel);
+
         return ninjaMapper.map(ninjaModel);
     }
 
     public List<NinjaDTO> listar(){
         List<NinjaModel> ninjaModelList = ninjaRepository.findAll();
+
         return ninjaModelList.stream().map(ninjaMapper::map).collect(Collectors.toList());
     }
 
     public NinjaDTO listarPorId(Long id){
         Optional<NinjaModel> optionalNinjaModel = ninjaRepository.findById(id);
+
         return optionalNinjaModel.map(ninjaMapper::map).orElse(null);
     }
 
     public NinjaDTO atualizaId(Long id, NinjaDTO ninjaDTO) {
-        Optional<NinjaModel> ninjaExistente = ninjaRepository.findById(id);
-        if (ninjaExistente.isPresent()){
-            NinjaModel ninjaAtualizado = ninjaMapper.map(ninjaDTO);
-            ninjaAtualizado.setId(id);
+        Optional<NinjaModel> optionalNinjaModel = ninjaRepository.findById(id);
+        if (optionalNinjaModel.isPresent()){
+            NinjaModel ninjaAtualizado = optionalNinjaModel.get();
+            ninjaMapper.updateModelFromDto(ninjaDTO,ninjaAtualizado);
             NinjaModel ninjaSalvo = ninjaRepository.save(ninjaAtualizado);
+
              return ninjaMapper.map(ninjaSalvo);
         }
         return null;
     }
 
-    public void deletarId(Long id) {
+    public boolean deletarId(Long id) {
+        if (!ninjaRepository.existsById(id)){
+            return false;
+        }
         ninjaRepository.deleteById(id);
+        return true;
     }
 }
